refactor(player): split oversized player and visualizer files
ProjectMVisualizer.kt (1175 LOC) -> 471 LOC. Each visualization style
moves to its own sibling file (Spectrum, Circular, Particles, Fluid,
etc.) so the orchestrator router stays small.

Player screens (video, audio, modern audio) move their bottom bars,
sheets, and dialogs into sibling files in the same package, bringing
each below the 800-LOC guardrail:

- MediaVideoPlayerScreen.kt: 1121 -> 484 LOC
- ModernAudioPlayerScreen.kt: 1068 -> 572 LOC
- MediaAudioPlayerScreen.kt:   864 -> 571 LOC
