+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/CenterControls.kt
@@ -1,0 +1,125 @@
+package com.universalmedialibrary.ui.media.player
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.theme.*
+
+@Composable
+internal fun CenterControls(
+    isPlaying: Boolean,
+    isBuffering: Boolean,
+    hasNext: Boolean,
+    hasPrevious: Boolean,
+    onPlayPause: () -> Unit,
+    onSkipPrevious: () -> Unit,
+    onSkipNext: () -> Unit,
+    onRewind: () -> Unit,
+    onFastForward: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(MediaSpacing.XL),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        // Previous
+        IconButton(
+            onClick = onSkipPrevious,
+            enabled = hasPrevious,
+            modifier = Modifier.size(48.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Default.SkipPrevious,
+                contentDescription = "Previous",
+                tint = if (hasPrevious) Color.White else Color.White.copy(alpha = 0.3f),
+                modifier = Modifier.size(36.dp)
+            )
+        }
+        
+        // Rewind 10s
+        IconButton(
+            onClick = onRewind,
+            modifier = Modifier.size(48.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Default.Replay10,
+                contentDescription = "Rewind 10 seconds",
+                tint = Color.White,
+                modifier = Modifier.size(36.dp)
+            )
+        }
+        
+        // Play/Pause
+        if (isBuffering) {
+            CircularProgressIndicator(
+                modifier = Modifier.size(72.dp),
+                color = Color.White,
+                strokeWidth = 3.dp
+            )
+        } else {
+            Surface(
+                shape = CircleShape,
+                color = Color.White.copy(alpha = 0.2f),
+                modifier = Modifier.size(72.dp),
+                onClick = onPlayPause
+            ) {
+                Icon(
+                    imageVector = if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
+                    contentDescription = if (isPlaying) "Pause" else "Play",
+                    tint = Color.White,
+                    modifier = Modifier.padding(16.dp)
+                )
+            }
+        }
+        
+        // Forward 10s
+        IconButton(
+            onClick = onFastForward,
+            modifier = Modifier.size(48.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Default.Forward10,
+                contentDescription = "Forward 10 seconds",
+                tint = Color.White,
+                modifier = Modifier.size(36.dp)
+            )
+        }
+        
+        // Next
+        IconButton(
+            onClick = onSkipNext,
+            enabled = hasNext,
+            modifier = Modifier.size(48.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Default.SkipNext,
+                contentDescription = "Next",
+                tint = if (hasNext) Color.White else Color.White.copy(alpha = 0.3f),
+                modifier = Modifier.size(36.dp)
+            )
+        }
+    }
+}
