+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/MediaAudioPlayerScreen.kt
@@ -490,314 +490,21 @@ fun MediaAudioPlayerScreen(
 // TOP BAR
 // =============================================================================
 
-@Composable
-private fun PlayerTopBar(
-    title: String,
-    subtitle: String?,
-    isCasting: Boolean,
-    onClose: () -> Unit,
-    onCast: () -> Unit,
-    onQueue: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.SM),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        IconButton(onClick = onClose) {
-            Icon(
-                imageVector = Icons.Default.ExpandMore,
-                contentDescription = "Minimize",
-                tint = MediaColors.TextPrimary
-            )
-        }
-        
-        Column(
-            modifier = Modifier.weight(1f),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = title,
-                style = MediaTypography.LabelMedium,
-                color = MediaColors.TextSecondary
-            )
-            subtitle?.let {
-                Text(
-                    text = it,
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextTertiary
-                )
-            }
-        }
-        
-        IconButton(onClick = onCast) {
-            Icon(
-                imageVector = if (isCasting) Icons.Filled.Cast else Icons.Outlined.Cast,
-                contentDescription = "Cast",
-                tint = if (isCasting) MediaColors.AccentPrimary else MediaColors.TextSecondary
-            )
-        }
-        
-        IconButton(onClick = onQueue) {
-            Icon(
-                imageVector = Icons.Default.MoreVert,
-                contentDescription = "More",
-                tint = MediaColors.TextSecondary
-            )
-        }
-    }
-}
 
 // =============================================================================
 // QUEUE SHEET
 // =============================================================================
 
-@Composable
-private fun QueueSheet(
-    queue: List<QueueItem>,
-    currentIndex: Int,
-    onItemClick: (Int) -> Unit,
-    onRemove: (Int) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MediaSpacing.MD)
-    ) {
-        Text(
-            text = "Up Next",
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            modifier = Modifier.padding(bottom = MediaSpacing.MD)
-        )
-        
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxWidth()
-                .heightIn(max = 400.dp)
-        ) {
-            itemsIndexed(queue) { index, item ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { onItemClick(index) }
-                        .background(
-                            if (index == currentIndex)
-                                MediaColors.AccentPrimary.copy(alpha = 0.1f)
-                            else
-                                Color.Transparent
-                        )
-                        .padding(MediaSpacing.SM),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    // Drag handle
-                    Icon(
-                        imageVector = Icons.Default.DragHandle,
-                        contentDescription = "Reorder",
-                        tint = MediaColors.TextTertiary,
-                        modifier = Modifier.size(20.dp)
-                    )
-                    
-                    Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                    
-                    // Artwork
-                    Surface(
-                        modifier = Modifier
-                            .size(48.dp)
-                            .clip(RoundedCornerShape(MediaCorners.XS)),
-                        color = MediaColors.BackgroundSurface
-                    ) {
-                        if (item.artworkUrl != null) {
-                            AsyncImage(
-                    
-                                model = item.artworkUrl,
-                                contentDescription = "Media image",
-                                contentScale = ContentScale.Crop,
-                                modifier = Modifier.fillMaxSize()
-                            )
-                        }
-                    }
-                    
-                    Spacer(modifier = Modifier.width(MediaSpacing.MD))
-                    
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = item.title,
-                            style = MediaTypography.BodyMedium,
-                            color = if (index == currentIndex) 
-                                MediaColors.AccentPrimary 
-                            else 
-                                MediaColors.TextPrimary,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            fontWeight = if (index == currentIndex) FontWeight.SemiBold else FontWeight.Normal
-                        )
-                        Text(
-                            text = item.artist ?: "",
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextSecondary,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis
-                        )
-                    }
-                    
-                    // Now playing indicator
-                    if (index == currentIndex) {
-                        Icon(
-                            imageVector = Icons.Default.VolumeUp,
-                            contentDescription = "Playing",
-                            tint = MediaColors.AccentPrimary,
-                            modifier = Modifier.size(20.dp)
-                        )
-                    } else {
-                        IconButton(onClick = { onRemove(index) }) {
-                            Icon(
-                                imageVector = Icons.Default.Close,
-                                contentDescription = "Remove",
-                                tint = MediaColors.TextTertiary,
-                                modifier = Modifier.size(20.dp)
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // SPEED PICKER
 // =============================================================================
 
-@Composable
-private fun SpeedPickerDialog(
-    currentSpeed: Float,
-    onSpeedChange: (Float) -> Unit,
-    onDismiss: () -> Unit
-) {
-    val speeds = listOf(0.5f, 0.75f, 1.0f, 1.25f, 1.5f, 1.75f, 2.0f, 2.5f, 3.0f)
-    
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text("Playback Speed", color = MediaColors.TextPrimary)
-        },
-        text = {
-            LazyColumn {
-                items(speeds) { speed ->
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { onSpeedChange(speed) }
-                            .padding(MediaSpacing.MD),
-                        horizontalArrangement = Arrangement.SpaceBetween,
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            text = "${speed}x",
-                            style = MediaTypography.BodyMedium,
-                            color = if (speed == currentSpeed) 
-                                MediaColors.AccentPrimary 
-                            else 
-                                MediaColors.TextPrimary,
-                            fontWeight = if (speed == currentSpeed) FontWeight.SemiBold else FontWeight.Normal
-                        )
-                        if (speed == currentSpeed) {
-                            Icon(
-                                imageVector = Icons.Default.Check,
-                                contentDescription = "Media image",
-                                tint = MediaColors.AccentPrimary
-                            )
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {},
-        containerColor = MediaColors.BackgroundElevated
-    )
-}
 
 // =============================================================================
 // SLEEP TIMER
 // =============================================================================
 
-@Composable
-private fun SleepTimerDialog(
-    currentTimer: Long?,
-    isActive: Boolean,
-    onSetTimer: (Int) -> Unit,
-    onDismiss: () -> Unit
-) {
-    val options = listOf(
-        5 to "5 minutes",
-        15 to "15 minutes",
-        30 to "30 minutes",
-        45 to "45 minutes",
-        60 to "1 hour",
-        -1 to "End of chapter/episode"
-    )
-    
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text("Sleep Timer", color = MediaColors.TextPrimary)
-        },
-        text = {
-            Column {
-                if (isActive && currentTimer != null) {
-                    Surface(
-                        color = MediaColors.AccentPrimary.copy(alpha = 0.1f),
-                        shape = RoundedCornerShape(MediaCorners.SM),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(bottom = MediaSpacing.MD)
-                    ) {
-                        Row(
-                            modifier = Modifier.padding(MediaSpacing.MD),
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Icon(
-                                imageVector = Icons.Default.Timer,
-                                contentDescription = "Media image",
-                                tint = MediaColors.AccentPrimary
-                            )
-                            Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                            Text(
-                                text = "Timer: ${formatTime(currentTimer)}",
-                                color = MediaColors.AccentPrimary,
-                                style = MediaTypography.BodyMedium
-                            )
-                            Spacer(modifier = Modifier.weight(1f))
-                            TextButton(onClick = { onSetTimer(0) }) {
-                                Text("Cancel", color = MediaColors.Error)
-                            }
-                        }
-                    }
-                }
-                
-                options.forEach { (minutes, label) ->
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { onSetTimer(minutes) }
-                            .padding(MediaSpacing.MD),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Text(
-                            text = label,
-                            style = MediaTypography.BodyMedium,
-                            color = MediaColors.TextPrimary
-                        )
-                    }
-                }
-            }
-        },
-        confirmButton = {},
-        containerColor = MediaColors.BackgroundElevated
-    )
-}
 
 // =============================================================================
 // HELPERS
