+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/MediaVideoPlayerScreen.kt
@@ -362,670 +362,33 @@ fun MediaVideoPlayerScreen(
 // TOP BAR
 // =============================================================================
 
-@Composable
-private fun VideoPlayerTopBar(
-    title: String,
-    subtitle: String?,
-    isCasting: Boolean,
-    onClose: () -> Unit,
-    onCast: () -> Unit,
-    onPip: () -> Unit,
-    onSettings: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .statusBarsPadding()
-            .padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.SM),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        IconButton(onClick = onClose) {
-            Icon(
-                imageVector = Icons.Default.ArrowBack,
-                contentDescription = "Back",
-                tint = Color.White
-            )
-        }
-        
-        Column(
-            modifier = Modifier.weight(1f),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = title,
-                style = MediaTypography.BodyLarge,
-                color = Color.White,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = FontWeight.Medium
-            )
-            subtitle?.let {
-                Text(
-                    text = it,
-                    style = MediaTypography.LabelSmall,
-                    color = Color.White.copy(alpha = 0.7f),
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-            }
-        }
-        
-        Row {
-            IconButton(onClick = onCast) {
-                Icon(
-                    imageVector = if (isCasting) Icons.Filled.Cast else Icons.Outlined.Cast,
-                    contentDescription = "Cast",
-                    tint = if (isCasting) MediaColors.AccentPrimary else Color.White
-                )
-            }
-            
-            IconButton(onClick = onPip) {
-                Icon(
-                    imageVector = Icons.Default.PictureInPicture,
-                    contentDescription = "Picture in Picture",
-                    tint = Color.White
-                )
-            }
-            
-            IconButton(onClick = onSettings) {
-                Icon(
-                    imageVector = Icons.Default.Settings,
-                    contentDescription = "Settings",
-                    tint = Color.White
-                )
-            }
-        }
-    }
-}
 
 // =============================================================================
 // CENTER CONTROLS
 // =============================================================================
 
-@Composable
-private fun CenterControls(
-    isPlaying: Boolean,
-    isBuffering: Boolean,
-    hasNext: Boolean,
-    hasPrevious: Boolean,
-    onPlayPause: () -> Unit,
-    onSkipPrevious: () -> Unit,
-    onSkipNext: () -> Unit,
-    onRewind: () -> Unit,
-    onFastForward: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.spacedBy(MediaSpacing.XL),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Previous
-        IconButton(
-            onClick = onSkipPrevious,
-            enabled = hasPrevious,
-            modifier = Modifier.size(48.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Default.SkipPrevious,
-                contentDescription = "Previous",
-                tint = if (hasPrevious) Color.White else Color.White.copy(alpha = 0.3f),
-                modifier = Modifier.size(36.dp)
-            )
-        }
-        
-        // Rewind 10s
-        IconButton(
-            onClick = onRewind,
-            modifier = Modifier.size(48.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Default.Replay10,
-                contentDescription = "Rewind 10 seconds",
-                tint = Color.White,
-                modifier = Modifier.size(36.dp)
-            )
-        }
-        
-        // Play/Pause
-        if (isBuffering) {
-            CircularProgressIndicator(
-                modifier = Modifier.size(72.dp),
-                color = Color.White,
-                strokeWidth = 3.dp
-            )
-        } else {
-            Surface(
-                shape = CircleShape,
-                color = Color.White.copy(alpha = 0.2f),
-                modifier = Modifier.size(72.dp),
-                onClick = onPlayPause
-            ) {
-                Icon(
-                    imageVector = if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
-                    contentDescription = if (isPlaying) "Pause" else "Play",
-                    tint = Color.White,
-                    modifier = Modifier.padding(16.dp)
-                )
-            }
-        }
-        
-        // Forward 10s
-        IconButton(
-            onClick = onFastForward,
-            modifier = Modifier.size(48.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Default.Forward10,
-                contentDescription = "Forward 10 seconds",
-                tint = Color.White,
-                modifier = Modifier.size(36.dp)
-            )
-        }
-        
-        // Next
-        IconButton(
-            onClick = onSkipNext,
-            enabled = hasNext,
-            modifier = Modifier.size(48.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Default.SkipNext,
-                contentDescription = "Next",
-                tint = if (hasNext) Color.White else Color.White.copy(alpha = 0.3f),
-                modifier = Modifier.size(36.dp)
-            )
-        }
-    }
-}
 
 // =============================================================================
 // BOTTOM BAR
 // =============================================================================
 
-@Composable
-private fun VideoPlayerBottomBar(
-    currentPosition: Long,
-    duration: Long,
-    bufferedPosition: Long,
-    chapters: List<VideoChapter>,
-    onSeek: (Long) -> Unit,
-    onSubtitles: () -> Unit,
-    onQuality: () -> Unit,
-    onEpisodes: (() -> Unit)?,
-    hasSubtitles: Boolean,
-    currentSubtitle: SubtitleTrack?,
-    currentQuality: VideoQuality,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .navigationBarsPadding()
-            .padding(MediaSpacing.MD)
-    ) {
-        // Progress bar with chapters
-        Box(modifier = Modifier.fillMaxWidth()) {
-            // Buffered progress
-            LinearProgressIndicator(
-                progress = { if (duration > 0) bufferedPosition.toFloat() / duration else 0f },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(4.dp)
-                    .clip(RoundedCornerShape(MediaCorners.Full)),
-                color = Color.White.copy(alpha = 0.3f),
-                trackColor = Color.White.copy(alpha = 0.1f)
-            )
-            
-            // Playback progress
-            Slider(
-                value = currentPosition.toFloat(),
-                onValueChange = { onSeek(it.toLong()) },
-                valueRange = 0f..duration.toFloat(),
-                modifier = Modifier.fillMaxWidth(),
-                colors = SliderDefaults.colors(
-                    thumbColor = MediaColors.AccentPrimary,
-                    activeTrackColor = MediaColors.AccentPrimary,
-                    inactiveTrackColor = Color.Transparent
-                )
-            )
-            
-            // Chapter markers
-            chapters.forEach { chapter ->
-                val position = chapter.startPosition.toFloat() / duration
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth(position)
-                        .height(4.dp)
-                        .align(Alignment.CenterStart)
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .align(Alignment.CenterEnd)
-                            .size(4.dp)
-                            .background(Color.White, CircleShape)
-                    )
-                }
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.XS))
-        
-        // Time and controls
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Time
-            Text(
-                text = "${formatTime(currentPosition)} / ${formatTime(duration)}",
-                style = MediaTypography.LabelMedium,
-                color = Color.White
-            )
-            
-            // Controls
-            Row(horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)) {
-                // Episodes (for TV shows)
-                onEpisodes?.let {
-                    TextButton(onClick = it) {
-                        Icon(
-                            imageVector = Icons.Default.VideoLibrary,
-                            contentDescription = "Media image",
-                            modifier = Modifier.size(20.dp)
-                        )
-                        Spacer(modifier = Modifier.width(MediaSpacing.XS))
-                        Text("Episodes")
-                    }
-                }
-                
-                // Subtitles
-                if (hasSubtitles) {
-                    TextButton(onClick = onSubtitles) {
-                        Icon(
-                            imageVector = if (currentSubtitle != null) 
-                                Icons.Filled.Subtitles 
-                            else 
-                                Icons.Outlined.Subtitles,
-                            contentDescription = "Media image",
-                            modifier = Modifier.size(20.dp),
-                            tint = if (currentSubtitle != null) 
-                                MediaColors.AccentPrimary 
-                            else 
-                                Color.White
-                        )
-                        Spacer(modifier = Modifier.width(MediaSpacing.XS))
-                        Text(
-                            text = currentSubtitle?.language ?: "Subtitles",
-                            color = if (currentSubtitle != null) 
-                                MediaColors.AccentPrimary 
-                            else 
-                                Color.White
-                        )
-                    }
-                }
-                
-                // Quality
-                TextButton(onClick = onQuality) {
-                    Icon(
-                        imageVector = Icons.Default.HighQuality,
-                        contentDescription = "Media image",
-                        modifier = Modifier.size(20.dp)
-                    )
-                    Spacer(modifier = Modifier.width(MediaSpacing.XS))
-                    Text(currentQuality.label)
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // VERTICAL INDICATOR
 // =============================================================================
 
-@Composable
-private fun VerticalIndicator(
-    icon: androidx.compose.ui.graphics.vector.ImageVector,
-    level: Float
-) {
-    Surface(
-        shape = RoundedCornerShape(MediaCorners.MD),
-        color = Color.Black.copy(alpha = 0.7f)
-    ) {
-        Column(
-            modifier = Modifier.padding(MediaSpacing.MD),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Icon(
-                imageVector = icon,
-                contentDescription = "Media image",
-                tint = Color.White,
-                modifier = Modifier.size(24.dp)
-            )
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.SM))
-            
-            Box(
-                modifier = Modifier
-                    .width(4.dp)
-                    .height(100.dp)
-                    .clip(RoundedCornerShape(MediaCorners.Full))
-                    .background(Color.White.copy(alpha = 0.3f))
-            ) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .fillMaxHeight(level)
-                        .align(Alignment.BottomCenter)
-                        .background(Color.White, RoundedCornerShape(MediaCorners.Full))
-                )
-            }
-        }
-    }
-}
 
 // =============================================================================
 // SEEK PREVIEW
 // =============================================================================
 
-@Composable
-private fun SeekPreview(
-    position: Long,
-    thumbnailUrl: String?,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Surface(
-            modifier = Modifier
-                .size(width = 160.dp, height = 90.dp)
-                .clip(RoundedCornerShape(MediaCorners.SM)),
-            color = Color.Black
-        ) {
-            if (thumbnailUrl != null) {
-                AsyncImage(
-                    
-                    model = thumbnailUrl,
-                    contentDescription = "Media image",
-                    contentScale = ContentScale.Crop,
-                    modifier = Modifier.fillMaxSize()
-                )
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(MediaSpacing.XS))
-        
-        Surface(
-            shape = RoundedCornerShape(MediaCorners.XS),
-            color = Color.Black.copy(alpha = 0.8f)
-        ) {
-            Text(
-                text = formatTime(position),
-                style = MediaTypography.LabelMedium,
-                color = Color.White,
-                modifier = Modifier.padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.XS)
-            )
-        }
-    }
-}
 
 // =============================================================================
 // SHEETS
 // =============================================================================
 
-@Composable
-private fun SubtitlesSheet(
-    tracks: List<SubtitleTrack>,
-    currentTrack: SubtitleTrack?,
-    onSelect: (SubtitleTrack?) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MediaSpacing.MD)
-    ) {
-        Text(
-            text = "Subtitles",
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            modifier = Modifier.padding(bottom = MediaSpacing.MD)
-        )
-        
-        // Off option
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { onSelect(null) }
-                .padding(MediaSpacing.MD),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = "Off",
-                style = MediaTypography.BodyMedium,
-                color = if (currentTrack == null) MediaColors.AccentPrimary else MediaColors.TextPrimary
-            )
-            if (currentTrack == null) {
-                Icon(
-                    imageVector = Icons.Default.Check,
-                    contentDescription = "Media image",
-                    tint = MediaColors.AccentPrimary
-                )
-            }
-        }
-        
-        HorizontalDivider(color = MediaColors.Border)
-        
-        tracks.forEach { track ->
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clickable { onSelect(track) }
-                    .padding(MediaSpacing.MD),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Column {
-                    Text(
-                        text = track.language,
-                        style = MediaTypography.BodyMedium,
-                        color = if (currentTrack?.id == track.id) 
-                            MediaColors.AccentPrimary 
-                        else 
-                            MediaColors.TextPrimary
-                    )
-                    track.label?.let {
-                        Text(
-                            text = it,
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextSecondary
-                        )
-                    }
-                }
-                if (currentTrack?.id == track.id) {
-                    Icon(
-                        imageVector = Icons.Default.Check,
-                        contentDescription = "Media image",
-                        tint = MediaColors.AccentPrimary
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun QualitySheet(
-    qualities: List<VideoQuality>,
-    currentQuality: VideoQuality,
-    onSelect: (VideoQuality) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MediaSpacing.MD)
-    ) {
-        Text(
-            text = "Quality",
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            modifier = Modifier.padding(bottom = MediaSpacing.MD)
-        )
-        
-        qualities.forEach { quality ->
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clickable { onSelect(quality) }
-                    .padding(MediaSpacing.MD),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Text(
-                        text = quality.label,
-                        style = MediaTypography.BodyMedium,
-                        color = if (currentQuality == quality) 
-                            MediaColors.AccentPrimary 
-                        else 
-                            MediaColors.TextPrimary
-                    )
-                    quality.bitrate?.let { bitrate ->
-                        Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                        Text(
-                            text = bitrate,
-                            style = MediaTypography.LabelSmall,
-                            color = MediaColors.TextTertiary
-                        )
-                    }
-                }
-                if (currentQuality == quality) {
-                    Icon(
-                        imageVector = Icons.Default.Check,
-                        contentDescription = "Media image",
-                        tint = MediaColors.AccentPrimary
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun EpisodesSheet(
-    episodes: List<EpisodeInfo>,
-    currentEpisode: String?,
-    onSelect: (EpisodeInfo) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MediaSpacing.MD)
-    ) {
-        Text(
-            text = "Episodes",
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            modifier = Modifier.padding(bottom = MediaSpacing.MD)
-        )
-        
-        LazyColumn(
-            modifier = Modifier.heightIn(max = 400.dp)
-        ) {
-            items(episodes) { episode ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable { onSelect(episode) }
-                        .background(
-                            if (episode.id == currentEpisode)
-                                MediaColors.AccentPrimary.copy(alpha = 0.1f)
-                            else
-                                Color.Transparent
-                        )
-                        .padding(MediaSpacing.SM),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    // Thumbnail
-                    Surface(
-                        modifier = Modifier
-                            .size(width = 120.dp, height = 68.dp)
-                            .clip(RoundedCornerShape(MediaCorners.XS)),
-                        color = MediaColors.BackgroundSurface
-                    ) {
-                        if (episode.thumbnailUrl != null) {
-                            AsyncImage(
-                    
-                                model = episode.thumbnailUrl,
-                                contentDescription = "Media image",
-                                contentScale = ContentScale.Crop,
-                                modifier = Modifier.fillMaxSize()
-                            )
-                        }
-                        
-                        // Play indicator for current
-                        if (episode.id == currentEpisode) {
-                            Box(
-                                modifier = Modifier
-                                    .fillMaxSize()
-                                    .background(Color.Black.copy(alpha = 0.5f)),
-                                contentAlignment = Alignment.Center
-                            ) {
-                                Icon(
-                                    imageVector = Icons.Default.VolumeUp,
-                                    contentDescription = "Playing",
-                                    tint = Color.White
-                                )
-                            }
-                        }
-                    }
-                    
-                    Spacer(modifier = Modifier.width(MediaSpacing.MD))
-                    
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = "E${episode.number}: ${episode.title}",
-                            style = MediaTypography.BodyMedium,
-                            color = if (episode.id == currentEpisode)
-                                MediaColors.AccentPrimary
-                            else
-                                MediaColors.TextPrimary,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            fontWeight = if (episode.id == currentEpisode)
-                                FontWeight.SemiBold
-                            else
-                                FontWeight.Normal
-                        )
-                        
-                        Row(
-                            horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
-                        ) {
-                            Text(
-                                text = episode.duration,
-                                style = MediaTypography.LabelSmall,
-                                color = MediaColors.TextSecondary
-                            )
-                            
-                            if (episode.progress > 0) {
-                                Text("•", color = MediaColors.TextTertiary)
-                                Text(
-                                    text = "${(episode.progress * 100).toInt()}% watched",
-                                    style = MediaTypography.LabelSmall,
-                                    color = MediaColors.AccentPrimary
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // HELPERS
