+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/QualitySheet.kt
@@ -1,0 +1,85 @@
+package com.universalmedialibrary.ui.media.player
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.theme.*
+
+@Composable
+internal fun QualitySheet(
+    qualities: List<VideoQuality>,
+    currentQuality: VideoQuality,
+    onSelect: (VideoQuality) -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(MediaSpacing.MD)
+    ) {
+        Text(
+            text = "Quality",
+            style = MediaTypography.TitleMedium,
+            color = MediaColors.TextPrimary,
+            modifier = Modifier.padding(bottom = MediaSpacing.MD)
+        )
+        
+        qualities.forEach { quality ->
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable { onSelect(quality) }
+                    .padding(MediaSpacing.MD),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Text(
+                        text = quality.label,
+                        style = MediaTypography.BodyMedium,
+                        color = if (currentQuality == quality) 
+                            MediaColors.AccentPrimary 
+                        else 
+                            MediaColors.TextPrimary
+                    )
+                    quality.bitrate?.let { bitrate ->
+                        Spacer(modifier = Modifier.width(MediaSpacing.SM))
+                        Text(
+                            text = bitrate,
+                            style = MediaTypography.LabelSmall,
+                            color = MediaColors.TextTertiary
+                        )
+                    }
+                }
+                if (currentQuality == quality) {
+                    Icon(
+                        imageVector = Icons.Default.Check,
+                        contentDescription = "Media image",
+                        tint = MediaColors.AccentPrimary
+                    )
+                }
+            }
+        }
+    }
+}
