+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/SeekPreview.kt
@@ -1,0 +1,71 @@
+package com.universalmedialibrary.ui.media.player
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.theme.*
+
+@Composable
+internal fun SeekPreview(
+    position: Long,
+    thumbnailUrl: String?,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Surface(
+            modifier = Modifier
+                .size(width = 160.dp, height = 90.dp)
+                .clip(RoundedCornerShape(MediaCorners.SM)),
+            color = Color.Black
+        ) {
+            if (thumbnailUrl != null) {
+                AsyncImage(
+                    
+                    model = thumbnailUrl,
+                    contentDescription = "Media image",
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier.fillMaxSize()
+                )
+            }
+        }
+        
+        Spacer(modifier = Modifier.height(MediaSpacing.XS))
+        
+        Surface(
+            shape = RoundedCornerShape(MediaCorners.XS),
+            color = Color.Black.copy(alpha = 0.8f)
+        ) {
+            Text(
+                text = formatTime(position),
+                style = MediaTypography.LabelMedium,
+                color = Color.White,
+                modifier = Modifier.padding(horizontal = MediaSpacing.SM, vertical = MediaSpacing.XS)
+            )
+        }
+    }
+}
