+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/player/VerticalIndicator.kt
@@ -1,0 +1,69 @@
+package com.universalmedialibrary.ui.media.player
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.gestures.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.AndroidView
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.theme.*
+
+@Composable
+internal fun VerticalIndicator(
+    icon: androidx.compose.ui.graphics.vector.ImageVector,
+    level: Float
+) {
+    Surface(
+        shape = RoundedCornerShape(MediaCorners.MD),
+        color = Color.Black.copy(alpha = 0.7f)
+    ) {
+        Column(
+            modifier = Modifier.padding(MediaSpacing.MD),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Icon(
+                imageVector = icon,
+                contentDescription = "Media image",
+                tint = Color.White,
+                modifier = Modifier.size(24.dp)
+            )
+            
+            Spacer(modifier = Modifier.height(MediaSpacing.SM))
+            
+            Box(
+                modifier = Modifier
+                    .width(4.dp)
+                    .height(100.dp)
+                    .clip(RoundedCornerShape(MediaCorners.Full))
+                    .background(Color.White.copy(alpha = 0.3f))
+            ) {
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .fillMaxHeight(level)
+                        .align(Alignment.BottomCenter)
+                        .background(Color.White, RoundedCornerShape(MediaCorners.Full))
+                )
+            }
+        }
+    }
+}
