+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/player/ModernAudioPlayerScreen.kt
@@ -512,110 +512,6 @@ fun ModernAudioPlayerScreen(
     }
 }
 
-@Composable
-private fun PlaybackControls(
-    isPlaying: Boolean,
-    onPlayPause: () -> Unit,
-    onPrevious: () -> Unit,
-    onNext: () -> Unit,
-    onShuffle: () -> Unit,
-    onRepeat: () -> Unit,
-    isShuffleEnabled: Boolean,
-    repeatMode: RepeatMode,
-    iconTint: Color,
-    highlightColor: Color
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Shuffle
-        IconButton(onClick = onShuffle) {
-            Icon(
-                Icons.Default.Shuffle,
-                contentDescription = "Shuffle",
-                tint = if (isShuffleEnabled) highlightColor else iconTint.copy(alpha = 0.5f),
-                modifier = Modifier.size(28.dp)
-            )
-        }
-
-        // Previous
-        Surface(
-            shape = CircleShape,
-            color = Color.White.copy(alpha = 0.1f),
-            modifier = Modifier.size(64.dp)
-        ) {
-            IconButton(onClick = onPrevious) {
-                Icon(
-                    Icons.Default.SkipPrevious,
-                    contentDescription = "Previous",
-                    tint = iconTint,
-                    modifier = Modifier.size(32.dp)
-                )
-            }
-        }
-
-        // Play/Pause
-        val scale by animateFloatAsState(
-            targetValue = if (isPlaying) 1f else 0.95f,
-            animationSpec = tween(200),
-            label = "playPauseScale"
-        )
-
-        Surface(
-            shape = CircleShape,
-            color = Color.White,
-            modifier = Modifier
-                .size(80.dp)
-                .scale(scale),
-            tonalElevation = 8.dp,
-            shadowElevation = 16.dp
-        ) {
-            IconButton(onClick = onPlayPause) {
-                Icon(
-                    if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
-                    contentDescription = if (isPlaying) "Pause" else "Play",
-                    tint = Color(0xFF1A1A2E),
-                    modifier = Modifier.size(40.dp)
-                )
-            }
-        }
-
-        // Next
-        Surface(
-            shape = CircleShape,
-            color = Color.White.copy(alpha = 0.1f),
-            modifier = Modifier.size(64.dp)
-        ) {
-            IconButton(onClick = onNext) {
-                Icon(
-                    Icons.Default.SkipNext,
-                    contentDescription = "Next",
-                    tint = iconTint,
-                    modifier = Modifier.size(32.dp)
-                )
-            }
-        }
-
-        // Repeat
-        IconButton(onClick = onRepeat) {
-            Icon(
-                when (repeatMode) {
-                    RepeatMode.OFF -> Icons.Default.Repeat
-                    RepeatMode.ALL -> Icons.Default.Repeat
-                    RepeatMode.ONE -> Icons.Default.RepeatOne
-                },
-                contentDescription = "Repeat",
-                tint = when (repeatMode) {
-                    RepeatMode.OFF -> iconTint.copy(alpha = 0.5f)
-                    else -> highlightColor
-                },
-                modifier = Modifier.size(28.dp)
-            )
-        }
-    }
-}
 
 @Composable
 private fun VinylRecord(
@@ -666,403 +562,11 @@ private fun formatOffset(offsetMs: Int): String {
     }
 }
 
-@Composable
-private fun WaveformSeekBar(
-    points: List<Float>,
-    progress: Float,
-    accent: Color,
-    backgroundColor: Color,
-    onSeek: (Float) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Canvas(
-        modifier = modifier.pointerInput(points) {
-            detectTapGestures { offset ->
-                val fraction = (offset.x / size.width).coerceIn(0f, 1f)
-                onSeek(fraction)
-            }
-        }
-    ) {
-        val barSpacing = size.width / (points.size * 1.2f)
-        val barWidth = barSpacing * 0.6f
-        val centerY = size.height / 2f
-        val progressX = size.width * progress.coerceIn(0f, 1f)
 
-        points.forEachIndexed { index, amplitude ->
-            val x = index * barSpacing + barSpacing / 2
-            val height = amplitude * size.height / 2
-            val color = if (x <= progressX) accent else backgroundColor
-            drawRoundRect(
-                color = color,
-                topLeft = Offset(x - barWidth / 2, centerY - height),
-                size = androidx.compose.ui.geometry.Size(barWidth, height * 2),
-                cornerRadius = androidx.compose.ui.geometry.CornerRadius(barWidth / 2, barWidth / 2)
-            )
-        }
-    }
-}
 
-@Composable
-private fun WaveformAlignmentControls(
-    offsetMs: Int,
-    accent: Color,
-    textColor: Color,
-    onAdjust: (Int) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 4.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text(
-            text = "Waveform offset ${formatOffset(offsetMs)}",
-            style = MaterialTheme.typography.bodySmall,
-            color = textColor
-        )
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            listOf(-100, -25, 25, 100).forEach { delta ->
-                OutlinedButton(
-                    onClick = { onAdjust(delta) },
-                    border = BorderStroke(1.dp, accent.copy(alpha = 0.6f)),
-                    contentPadding = PaddingValues(horizontal = 12.dp, vertical = 4.dp)
-                ) {
-                    Text(
-                        text = "${if (delta > 0) "+" else ""}$delta ms",
-                        color = textColor,
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun QuickToggleRow(
-    skipSilenceEnabled: Boolean,
-    crossfadeDurationMs: Int,
-    accent: Color,
-    textColor: Color,
-    onSkipSilence: () -> Unit,
-    onCrossfadeToggle: () -> Unit,
-    onCrossfadeLongPress: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 8.dp),
-        horizontalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        QuickToggleChip(
-            label = if (skipSilenceEnabled) "Skip Silence On" else "Skip Silence",
-            enabled = skipSilenceEnabled,
-            accent = accent,
-            textColor = textColor,
-            onToggle = onSkipSilence
-        )
-        QuickToggleChip(
-            label = if (crossfadeDurationMs > 0) "Crossfade ${crossfadeDurationMs / 1000}s" else "Crossfade",
-            enabled = crossfadeDurationMs > 0,
-            accent = accent,
-            textColor = textColor,
-            onToggle = onCrossfadeToggle,
-            onLongPress = onCrossfadeLongPress
-        )
-    }
-}
 
-@Composable
-private fun RowScope.QuickToggleChip(
-    label: String,
-    enabled: Boolean,
-    accent: Color,
-    textColor: Color,
-    onToggle: () -> Unit,
-    onLongPress: (() -> Unit)? = null
-) {
-    val background = if (enabled) accent.copy(alpha = 0.2f) else Color.White.copy(alpha = 0.1f)
-    Surface(
-        shape = RoundedCornerShape(50),
-        color = background,
-        border = if (enabled) BorderStroke(1.dp, accent) else null,
-        modifier = Modifier
-            .weight(1f)
-            .pointerInput(label, enabled) {
-                detectTapGestures(
-                    onTap = { onToggle() },
-                    onLongPress = { onLongPress?.invoke() }
-                )
-            }
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp, vertical = 10.dp),
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(10.dp)
-                    .clip(CircleShape)
-                    .background(if (enabled) accent else textColor.copy(alpha = 0.4f))
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(label, color = textColor, style = MaterialTheme.typography.labelMedium)
-        }
-    }
-}
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun AdvancedPlaybackSheet(
-    skipSilenceEnabled: Boolean,
-    onSkipSilenceToggled: () -> Unit,
-    crossfadeSeconds: Int,
-    onCrossfadeDurationChanged: (Int) -> Unit,
-    onCrossfadeToggle: () -> Unit,
-    partyModeEnabled: Boolean,
-    onPartyModeToggle: () -> Unit,
-    onSleepTimerSelected: (Int) -> Unit,
-    onSleepTimerClear: () -> Unit,
-    lastSleepTimerMinutes: Int,
-    onDismiss: () -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .navigationBarsPadding()
-            .padding(24.dp),
-        verticalArrangement = Arrangement.spacedBy(20.dp)
-    ) {
-        Text("Playback Enhancements", style = MaterialTheme.typography.titleMedium)
 
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Column {
-                Text("Skip silence", style = MaterialTheme.typography.bodyLarge)
-                Text("Trim quiet gaps automatically", style = MaterialTheme.typography.bodySmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-            }
-            Switch(checked = skipSilenceEnabled, onCheckedChange = { onSkipSilenceToggled() })
-        }
 
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Column {
-                    Text("Crossfade", style = MaterialTheme.typography.bodyLarge)
-                    Text("Blend transitions between tracks", style = MaterialTheme.typography.bodySmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-                }
-                Switch(checked = crossfadeSeconds > 0, onCheckedChange = { onCrossfadeToggle() })
-            }
-            Slider(
-                value = crossfadeSeconds.toFloat(),
-                onValueChange = { onCrossfadeDurationChanged(it.roundToInt()) },
-                valueRange = 0f..12f,
-                steps = 11,
-                colors = SliderDefaults.colors(activeTrackColor = MaterialTheme.colorScheme.primary)
-            )
-            Text("${crossfadeSeconds}s", style = MaterialTheme.typography.labelMedium)
-        }
-
-        Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            Text("Sleep timer", style = MaterialTheme.typography.bodyLarge)
-            Row(
-                modifier = Modifier
-                    .horizontalScroll(rememberScrollState()),
-                horizontalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                listOf(15, 30, 45, 60).forEach { minutes ->
-                    FilterChip(
-                        selected = lastSleepTimerMinutes == minutes,
-                        onClick = { onSleepTimerSelected(minutes) },
-                        label = { Text("${minutes}m") },
-                        leadingIcon = if (lastSleepTimerMinutes == minutes) {
-                            { Icon(Icons.Default.Check, contentDescription = "Confirm", modifier = Modifier.size(16.dp)) }
-                        } else null
-                    )
-                }
-                FilterChip(
-                    selected = false,
-                    onClick = onSleepTimerClear,
-                    label = { Text("Clear") }
-                )
-            }
-        }
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Column {
-                Text("Party mode", style = MaterialTheme.typography.bodyLarge)
-                Text("Animated backdrop for celebrations", style = MaterialTheme.typography.bodySmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-            }
-            Switch(checked = partyModeEnabled, onCheckedChange = { onPartyModeToggle() })
-        }
-
-        Spacer(modifier = Modifier.height(8.dp))
-        Button(
-            onClick = onDismiss,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text("Done")
-        }
-    }
-}
-
-@Composable
-private fun QueueSheet(
-    entries: List<AudioQueueEntry>,
-    onPlay: (Int) -> Unit,
-    onMoveUp: (Int) -> Unit,
-    onMoveDown: (Int) -> Unit,
-    onRemove: (Int) -> Unit,
-    onClose: () -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .navigationBarsPadding()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        Text("Play Queue", style = MaterialTheme.typography.titleMedium)
-        if (entries.isEmpty()) {
-            Text("Queue is empty", color = MaterialTheme.colorScheme.onSurfaceVariant)
-        } else {
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-                items(entries, key = { it.index }) { entry ->
-                    Surface(
-                        shape = RoundedCornerShape(12.dp),
-                        tonalElevation = if (entry.isCurrent) 4.dp else 0.dp,
-                        color = if (entry.isCurrent) MaterialTheme.colorScheme.primary.copy(alpha = 0.08f) else MaterialTheme.colorScheme.surfaceVariant
-                    ) {
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(12.dp),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Column(modifier = Modifier.weight(1f)) {
-                                Text(entry.title, style = MaterialTheme.typography.bodyLarge)
-                                if (entry.isCurrent) {
-                                    Text("Now playing", style = MaterialTheme.typography.labelSmall, color = MaterialTheme.colorScheme.primary)
-                                }
-                            }
-                            Row(horizontalArrangement = Arrangement.spacedBy(4.dp)) {
-                                IconButton(onClick = { onPlay(entry.index) }) {
-                                    Icon(Icons.Default.PlayArrow, contentDescription = "Play this track")
-                                }
-                                IconButton(onClick = { onMoveUp(entry.index) }, enabled = entry.index > 0) {
-                                    Icon(Icons.Default.KeyboardArrowUp, contentDescription = "Move up")
-                                }
-                                IconButton(onClick = { onMoveDown(entry.index) }, enabled = entry.index < entries.lastIndex) {
-                                    Icon(Icons.Default.KeyboardArrowDown, contentDescription = "Move down")
-                                }
-                                IconButton(onClick = { onRemove(entry.index) }) {
-                                    Icon(Icons.Default.Delete, contentDescription = "Remove")
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Button(
-            onClick = onClose,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Text("Close")
-        }
-    }
-}
-
-@Composable
-private fun SleepTimerBadge(
-    endTime: Long,
-    onClear: () -> Unit,
-    textColor: Color
-) {
-    var now by remember(endTime) { mutableStateOf(System.currentTimeMillis()) }
-    LaunchedEffect(endTime) {
-        while (now < endTime) {
-            delay(1000)
-            now = System.currentTimeMillis()
-        }
-    }
-    val remaining = (endTime - now).coerceAtLeast(0L)
-    if (remaining <= 0L) return
-    val minutes = (remaining / 60000)
-    val seconds = (remaining / 1000) % 60
-    Surface(
-        shape = RoundedCornerShape(50),
-        color = textColor.copy(alpha = 0.15f),
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = "Sleep timer ${String.format(Locale.US, "%02d:%02d", minutes, seconds)}",
-                color = textColor,
-                style = MaterialTheme.typography.labelLarge
-            )
-            IconButton(onClick = onClear) {
-                Icon(Icons.Default.Close, contentDescription = "Cancel timer", tint = textColor)
-            }
-        }
-    }
-}
-
-@Composable
-private fun PartyModeAurora(
-    modifier: Modifier = Modifier,
-    baseColors: List<Color>
-) {
-    val transition = rememberInfiniteTransition(label = "party_mode")
-    val phase by transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 1f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(durationMillis = 12000, easing = LinearEasing),
-            repeatMode = androidx.compose.animation.core.RepeatMode.Restart
-        ),
-        label = "party_phase"
-    )
-    Canvas(modifier = modifier) {
-        val width = size.width
-        val height = size.height
-        val palette = listOf(
-            baseColors.getOrNull(0) ?: Color.Magenta,
-            baseColors.getOrNull(1) ?: Color.Cyan,
-            baseColors.getOrNull(2) ?: Color.Yellow
-        )
-        palette.forEachIndexed { index, color ->
-            val animatedRadius = (width.coerceAtMost(height) / 2f) * (0.4f + 0.25f * sin((phase + index * 0.3f) * 2 * PI).toFloat())
-            val animatedX = width * ((phase + index * 0.2f) % 1f)
-            val animatedY = height * (((phase * 1.2f) + index * 0.15f) % 1f)
-            drawCircle(
-                color = color.copy(alpha = 0.12f),
-                radius = animatedRadius,
-                center = Offset(animatedX, animatedY),
-                style = Stroke(width = animatedRadius * 0.15f)
-            )
-        }
-    }
-}
