+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/player/PlaybackControls.kt
@@ -1,0 +1,156 @@
+package com.universalmedialibrary.ui.player
+
+import androidx.compose.animation.*
+import androidx.compose.animation.core.*
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.blur
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.graphics.Shadow
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import coil.compose.AsyncImage
+import com.universalmedialibrary.services.audio.AudioPlaybackManager.AudioQueueEntry
+import com.universalmedialibrary.ui.player.WaveformSource
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import java.util.Locale
+import kotlin.math.PI
+import kotlin.math.absoluteValue
+import kotlin.math.cos
+import kotlin.math.sin
+import kotlin.math.roundToInt
+
+@Composable
+internal fun PlaybackControls(
+    isPlaying: Boolean,
+    onPlayPause: () -> Unit,
+    onPrevious: () -> Unit,
+    onNext: () -> Unit,
+    onShuffle: () -> Unit,
+    onRepeat: () -> Unit,
+    isShuffleEnabled: Boolean,
+    repeatMode: RepeatMode,
+    iconTint: Color,
+    highlightColor: Color
+) {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.SpaceEvenly,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        // Shuffle
+        IconButton(onClick = onShuffle) {
+            Icon(
+                Icons.Default.Shuffle,
+                contentDescription = "Shuffle",
+                tint = if (isShuffleEnabled) highlightColor else iconTint.copy(alpha = 0.5f),
+                modifier = Modifier.size(28.dp)
+            )
+        }
+
+        // Previous
+        Surface(
+            shape = CircleShape,
+            color = Color.White.copy(alpha = 0.1f),
+            modifier = Modifier.size(64.dp)
+        ) {
+            IconButton(onClick = onPrevious) {
+                Icon(
+                    Icons.Default.SkipPrevious,
+                    contentDescription = "Previous",
+                    tint = iconTint,
+                    modifier = Modifier.size(32.dp)
+                )
+            }
+        }
+
+        // Play/Pause
+        val scale by animateFloatAsState(
+            targetValue = if (isPlaying) 1f else 0.95f,
+            animationSpec = tween(200),
+            label = "playPauseScale"
+        )
+
+        Surface(
+            shape = CircleShape,
+            color = Color.White,
+            modifier = Modifier
+                .size(80.dp)
+                .scale(scale),
+            tonalElevation = 8.dp,
+            shadowElevation = 16.dp
+        ) {
+            IconButton(onClick = onPlayPause) {
+                Icon(
+                    if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
+                    contentDescription = if (isPlaying) "Pause" else "Play",
+                    tint = Color(0xFF1A1A2E),
+                    modifier = Modifier.size(40.dp)
+                )
+            }
+        }
+
+        // Next
+        Surface(
+            shape = CircleShape,
+            color = Color.White.copy(alpha = 0.1f),
+            modifier = Modifier.size(64.dp)
+        ) {
+            IconButton(onClick = onNext) {
+                Icon(
+                    Icons.Default.SkipNext,
+                    contentDescription = "Next",
+                    tint = iconTint,
+                    modifier = Modifier.size(32.dp)
+                )
+            }
+        }
+
+        // Repeat
+        IconButton(onClick = onRepeat) {
+            Icon(
+                when (repeatMode) {
+                    RepeatMode.OFF -> Icons.Default.Repeat
+                    RepeatMode.ALL -> Icons.Default.Repeat
+                    RepeatMode.ONE -> Icons.Default.RepeatOne
+                },
+                contentDescription = "Repeat",
+                tint = when (repeatMode) {
+                    RepeatMode.OFF -> iconTint.copy(alpha = 0.5f)
+                    else -> highlightColor
+                },
+                modifier = Modifier.size(28.dp)
+            )
+        }
+    }
+}
