+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/player/QuickToggleRow.kt
@@ -1,0 +1,85 @@
+package com.universalmedialibrary.ui.player
+
+import androidx.compose.animation.*
+import androidx.compose.animation.core.*
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.blur
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.graphics.Shadow
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import coil.compose.AsyncImage
+import com.universalmedialibrary.services.audio.AudioPlaybackManager.AudioQueueEntry
+import com.universalmedialibrary.ui.player.WaveformSource
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import java.util.Locale
+import kotlin.math.PI
+import kotlin.math.absoluteValue
+import kotlin.math.cos
+import kotlin.math.sin
+import kotlin.math.roundToInt
+
+@Composable
+internal fun QuickToggleRow(
+    skipSilenceEnabled: Boolean,
+    crossfadeDurationMs: Int,
+    accent: Color,
+    textColor: Color,
+    onSkipSilence: () -> Unit,
+    onCrossfadeToggle: () -> Unit,
+    onCrossfadeLongPress: () -> Unit
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 8.dp),
+        horizontalArrangement = Arrangement.spacedBy(12.dp)
+    ) {
+        QuickToggleChip(
+            label = if (skipSilenceEnabled) "Skip Silence On" else "Skip Silence",
+            enabled = skipSilenceEnabled,
+            accent = accent,
+            textColor = textColor,
+            onToggle = onSkipSilence
+        )
+        QuickToggleChip(
+            label = if (crossfadeDurationMs > 0) "Crossfade ${crossfadeDurationMs / 1000}s" else "Crossfade",
+            enabled = crossfadeDurationMs > 0,
+            accent = accent,
+            textColor = textColor,
+            onToggle = onCrossfadeToggle,
+            onLongPress = onCrossfadeLongPress
+        )
+    }
+}
