+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/visualizer/ProjectMVisualizer.kt
@@ -182,187 +182,8 @@ private fun TunnelVisualizer(
     }
 }
 
-/**
- * Starfield visualization
- */
-@Composable
-private fun StarfieldVisualizer(
-    state: VisualizerState,
-    rotation: Float,
-    starColor: Color
-) {
-    val spectrum = state.frequencyBands.spectrum.ifEmpty { List(64) { 0f } }
-    val bass = state.frequencyBands.bass
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        
-        // Use spectrum to determine star brightness/size
-        spectrum.forEachIndexed { index, magnitude ->
-            // Random-ish position based on index
-            val angle = (index * 137.5f + rotation) * PI.toFloat() / 180f
-            val distance = (index * 10f + (state.timestamp % 1000)) % (min(size.width, size.height) / 2)
-            
-            val x = centerX + cos(angle) * distance
-            val y = centerY + sin(angle) * distance
-            
-            val size = 2f + magnitude * 10f + bass * 5f
-            
-            drawCircle(
-                color = starColor.copy(alpha = magnitude.coerceIn(0.2f, 1f)),
-                radius = size,
-                center = Offset(x, y)
-            )
-            
-            // Trailing effect for fast moving stars (bass kick)
-            if (bass > 0.5f) {
-                drawLine(
-                    color = starColor.copy(alpha = 0.3f),
-                    start = Offset(centerX, centerY),
-                    end = Offset(x, y),
-                    strokeWidth = 1f
-                )
-            }
-        }
-    }
-}
 
-/**
- * Fluid visualization
- */
-@Composable
-private fun FluidVisualizer(
-    state: VisualizerState,
-    color1: Color,
-    color2: Color,
-    color3: Color
-) {
-    val waveform = state.waveform.ifEmpty { List(128) { 0f } }
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        val radius = min(size.width, size.height) * 0.4f
-        
-        val path = Path()
-        val points = waveform.size
-        
-        for (i in 0 until points) {
-            val angle = (i.toFloat() / points) * 2 * PI.toFloat()
-            val amplitude = waveform[i] * 100f
-            val r = radius + amplitude
-            
-            val x = centerX + cos(angle) * r
-            val y = centerY + sin(angle) * r
-            
-            if (i == 0) path.moveTo(x, y)
-            else path.lineTo(x, y)
-        }
-        path.close()
-        
-        // Draw filled fluid
-        drawPath(
-            path = path,
-            brush = Brush.radialGradient(
-                colors = listOf(color1, color2, color3),
-                center = Offset(centerX, centerY)
-            )
-        )
-        
-        // Draw outline
-        drawPath(
-            path = path,
-            color = color1.copy(alpha = 0.8f),
-            style = Stroke(width = 3f)
-        )
-    }
-}
 
-/**
- * Spectrum bars visualization (classic frequency bars with neon glow effect)
- */
-@Composable
-private fun SpectrumBarsVisualizer(
-    state: VisualizerState,
-    primaryColor: Color,
-    secondaryColor: Color
-) {
-    val spectrum = state.frequencyBands.spectrum.ifEmpty { List(64) { 0f } }
-    val bass = state.frequencyBands.bass
-
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val barWidth = size.width / spectrum.size
-        val maxHeight = size.height
-        val gap = 2f
-
-        spectrum.forEachIndexed { index, magnitude ->
-            val barHeight = magnitude * maxHeight * 0.9f
-            val x = index * barWidth
-            val progress = index.toFloat() / spectrum.size
-            
-            // Calculate rainbow color based on position and intensity
-            val hue = (progress * 270f + bass * 90f) % 360f
-            val barColor = Color.hsv(hue, 0.9f, 0.95f + magnitude * 0.05f)
-            val glowColor = Color.hsv(hue, 0.7f, 1f)
-
-            // Draw glow effect (wider, semi-transparent bar behind)
-            if (magnitude > 0.1f) {
-                drawRect(
-                    color = glowColor.copy(alpha = 0.3f * magnitude),
-                    topLeft = Offset(x - gap, maxHeight - barHeight - gap * 2),
-                    size = androidx.compose.ui.geometry.Size(barWidth + gap * 2, barHeight + gap * 4)
-                )
-            }
-
-            // Draw main bar with gradient
-            drawRect(
-                brush = Brush.verticalGradient(
-                    colors = listOf(
-                        barColor,
-                        barColor.copy(alpha = 0.8f),
-                        androidx.compose.ui.graphics.lerp(barColor, primaryColor, 0.3f).copy(alpha = 0.6f)
-                    ),
-                    startY = maxHeight - barHeight,
-                    endY = maxHeight
-                ),
-                topLeft = Offset(x + gap, maxHeight - barHeight),
-                size = androidx.compose.ui.geometry.Size(barWidth - gap * 2, barHeight)
-            )
-            
-            // Draw top cap (bright white highlight for peaks)
-            if (magnitude > 0.2f) {
-                drawRect(
-                    color = Color.White.copy(alpha = 0.8f * magnitude),
-                    topLeft = Offset(x + gap, maxHeight - barHeight),
-                    size = androidx.compose.ui.geometry.Size(barWidth - gap * 2, 3f)
-                )
-            }
-        }
-        
-        // Draw reflection effect at bottom
-        spectrum.forEachIndexed { index, magnitude ->
-            val barHeight = magnitude * maxHeight * 0.15f // Shorter reflection
-            val x = index * barWidth
-            val progress = index.toFloat() / spectrum.size
-            val hue = (progress * 270f + bass * 90f) % 360f
-            val barColor = Color.hsv(hue, 0.9f, 0.95f)
-
-            drawRect(
-                brush = Brush.verticalGradient(
-                    colors = listOf(
-                        barColor.copy(alpha = 0.3f),
-                        barColor.copy(alpha = 0f)
-                    ),
-                    startY = maxHeight,
-                    endY = maxHeight + barHeight
-                ),
-                topLeft = Offset(x + gap, maxHeight),
-                size = androidx.compose.ui.geometry.Size(barWidth - gap * 2, barHeight)
-            )
-        }
-    }
-}
 
 /**
  * Waveform visualization (time-domain)
@@ -403,217 +224,8 @@ private fun WaveformVisualizer(
     }
 }
 
-/**
- * Circular visualization (radial spectrum with rainbow effect)
- */
-@Composable
-private fun CircularVisualizer(
-    state: VisualizerState,
-    rotation: Float
-) {
-    val spectrum = state.frequencyBands.spectrum.ifEmpty { List(64) { 0f } }
-    val bass = state.frequencyBands.bass
-    val mid = state.frequencyBands.mid
 
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        val baseRadius = min(size.width, size.height) * 0.15f
-        val pulseRadius = baseRadius * (1f + bass * 0.3f) // Pulse with bass
-        
-        // Draw center glow
-        drawCircle(
-            brush = Brush.radialGradient(
-                colors = listOf(
-                    VisualizerColors.neonCyan.copy(alpha = 0.4f * bass),
-                    VisualizerColors.neonMagenta.copy(alpha = 0.2f * bass),
-                    Color.Transparent
-                ),
-                center = Offset(centerX, centerY),
-                radius = baseRadius * 2f
-            ),
-            radius = baseRadius * 2f,
-            center = Offset(centerX, centerY)
-        )
-        
-        // Draw inner circle
-        drawCircle(
-            color = VisualizerColors.neonCyan.copy(alpha = 0.3f),
-            radius = pulseRadius * 0.5f,
-            center = Offset(centerX, centerY),
-            style = Stroke(width = 2f)
-        )
 
-        // Draw spectrum bars radiating outward
-        spectrum.forEachIndexed { index, magnitude ->
-            val angle = (index.toFloat() / spectrum.size * 360f + rotation) * PI.toFloat() / 180f
-            val length = magnitude * min(size.width, size.height) * 0.35f
-            val startRadius = pulseRadius
-            val endRadius = pulseRadius + length
-
-            val startX = centerX + cos(angle) * startRadius
-            val startY = centerY + sin(angle) * startRadius
-            val endX = centerX + cos(angle) * endRadius
-            val endY = centerY + sin(angle) * endRadius
-
-            // Rainbow color based on position
-            val hue = (index.toFloat() / spectrum.size * 360f + rotation * 0.5f) % 360f
-            val barColor = Color.hsv(hue, 0.9f, 0.95f)
-            val glowColor = Color.hsv(hue, 0.7f, 1f)
-
-            // Draw glow
-            if (magnitude > 0.15f) {
-                drawLine(
-                    color = glowColor.copy(alpha = 0.4f * magnitude),
-                    start = Offset(startX, startY),
-                    end = Offset(endX, endY),
-                    strokeWidth = 8f,
-                    cap = StrokeCap.Round
-                )
-            }
-
-            // Draw main line
-            drawLine(
-                brush = Brush.linearGradient(
-                    colors = listOf(
-                        barColor.copy(alpha = 0.7f),
-                        barColor,
-                        Color.White.copy(alpha = 0.8f * magnitude)
-                    ),
-                    start = Offset(startX, startY),
-                    end = Offset(endX, endY)
-                ),
-                start = Offset(startX, startY),
-                end = Offset(endX, endY),
-                strokeWidth = 3f + magnitude * 2f,
-                cap = StrokeCap.Round
-            )
-        }
-        
-        // Draw outer ring that pulses with mid frequencies
-        drawCircle(
-            color = VisualizerColors.neonMagenta.copy(alpha = 0.5f * mid),
-            radius = baseRadius + min(size.width, size.height) * 0.35f * mid,
-            center = Offset(centerX, centerY),
-            style = Stroke(width = 2f + mid * 3f)
-        )
-    }
-}
-
-/**
- * Particle visualization (reactive particles)
- */
-@Composable
-private fun ParticleVisualizer(
-    state: VisualizerState,
-    primaryColor: Color,
-    secondaryColor: Color,
-    tertiaryColor: Color
-) {
-    val bass = state.frequencyBands.bass
-    val mid = state.frequencyBands.mid
-    val treble = state.frequencyBands.treble
-
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-
-        // Bass particles (large)
-        drawCircle(
-            color = primaryColor.copy(alpha = bass),
-            radius = bass * 100f,
-            center = Offset(centerX, centerY)
-        )
-
-        // Mid particles (medium)
-        drawCircle(
-            color = secondaryColor.copy(alpha = mid),
-            radius = mid * 80f,
-            center = Offset(centerX - 100f, centerY)
-        )
-        drawCircle(
-            color = secondaryColor.copy(alpha = mid),
-            radius = mid * 80f,
-            center = Offset(centerX + 100f, centerY)
-        )
-
-        // Treble particles (small)
-        for (i in 0..7) {
-            val angle = (i / 8f) * 2 * PI.toFloat()
-            val x = centerX + cos(angle) * 150f
-            val y = centerY + sin(angle) * 150f
-            drawCircle(
-                color = tertiaryColor.copy(alpha = treble),
-                radius = treble * 40f,
-                center = Offset(x, y)
-            )
-        }
-    }
-}
-
-/**
- * Frequency rings visualization (concentric reactive rings)
- */
-@Composable
-private fun FrequencyRingsVisualizer(
-    state: VisualizerState,
-    rotation: Float,
-    primaryColor: Color,
-    secondaryColor: Color,
-    tertiaryColor: Color
-) {
-    val bass = state.frequencyBands.bass
-    val mid = state.frequencyBands.mid
-    val treble = state.frequencyBands.treble
-
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        val maxRadius = min(size.width, size.height) / 2
-
-        // Bass ring (outer)
-        drawCircle(
-            brush = Brush.radialGradient(
-                colors = listOf(
-                    primaryColor.copy(alpha = 0f),
-                    primaryColor.copy(alpha = bass * 0.8f)
-                ),
-                center = Offset(centerX, centerY)
-            ),
-            radius = maxRadius * 0.8f * (0.8f + bass * 0.2f),
-            center = Offset(centerX, centerY),
-            style = Stroke(width = 20f)
-        )
-
-        // Mid ring (middle)
-        drawCircle(
-            brush = Brush.radialGradient(
-                colors = listOf(
-                    secondaryColor.copy(alpha = 0f),
-                    secondaryColor.copy(alpha = mid * 0.8f)
-                ),
-                center = Offset(centerX, centerY)
-            ),
-            radius = maxRadius * 0.5f * (0.9f + mid * 0.1f),
-            center = Offset(centerX, centerY),
-            style = Stroke(width = 15f)
-        )
-
-        // Treble ring (inner)
-        drawCircle(
-            brush = Brush.radialGradient(
-                colors = listOf(
-                    tertiaryColor.copy(alpha = treble * 0.8f),
-                    tertiaryColor.copy(alpha = 0f)
-                ),
-                center = Offset(centerX, centerY)
-            ),
-            radius = maxRadius * 0.3f * (1.0f + treble * 0.2f),
-            center = Offset(centerX, centerY),
-            style = Stroke(width = 10f)
-        )
-    }
-}
 
 /**
  * Oscilloscope visualization (XY mode waveform)
@@ -655,50 +267,6 @@ private fun OscilloscopeVisualizer(
     }
 }
 
-/**
- * Spectrograph visualization (frequency over time)
- */
-@Composable
-private fun SpectrographVisualizer(
-    state: VisualizerState,
-    primaryColor: Color,
-    secondaryColor: Color,
-    tertiaryColor: Color
-) {
-    val spectrum = state.frequencyBands.spectrum.ifEmpty { List(64) { 0f } }
-    
-    // Store history for scrolling effect
-    val history = remember { mutableListOf<List<Float>>() }
-    LaunchedEffect(state.timestamp) {
-        history.add(spectrum)
-        if (history.size > 60) history.removeAt(0)
-    }
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val cellWidth = size.width / spectrum.size
-        val cellHeight = size.height / maxOf(history.size, 1)
-        
-        history.forEachIndexed { timeIndex, spectrumData ->
-            spectrumData.forEachIndexed { freqIndex, magnitude ->
-                val x = freqIndex * cellWidth
-                val y = timeIndex * cellHeight
-                val progress = magnitude
-                
-                val color = when {
-                    progress > 0.7f -> tertiaryColor
-                    progress > 0.4f -> secondaryColor
-                    else -> primaryColor
-                }.copy(alpha = progress.coerceIn(0.1f, 1f))
-                
-                drawRect(
-                    color = color,
-                    topLeft = Offset(x, y),
-                    size = androidx.compose.ui.geometry.Size(cellWidth * 0.9f, cellHeight * 0.9f)
-                )
-            }
-        }
-    }
-}
 
 /**
  * Lissajous curves visualization
@@ -788,81 +356,6 @@ private fun RadialWaveformVisualizer(
     }
 }
 
-/**
- * Beat-reactive shapes visualization
- */
-@Composable
-private fun BeatReactiveVisualizer(
-    state: VisualizerState,
-    primaryColor: Color,
-    secondaryColor: Color,
-    tertiaryColor: Color
-) {
-    val bass = state.frequencyBands.bass
-    val mid = state.frequencyBands.mid
-    val treble = state.frequencyBands.treble
-    val isBeat = state.isBeat
-    
-    // Animate beat pulse
-    val beatPulse = remember { Animatable(1f) }
-    LaunchedEffect(isBeat) {
-        if (isBeat) {
-            beatPulse.snapTo(1.5f)
-            beatPulse.animateTo(1f, animationSpec = tween(300, easing = FastOutSlowInEasing))
-        }
-    }
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        val pulseScale = beatPulse.value
-        
-        // Central circle reacts to beat
-        drawCircle(
-            color = primaryColor.copy(alpha = 0.8f),
-            radius = bass * 150f * pulseScale,
-            center = Offset(centerX, centerY)
-        )
-        
-        // Rotating triangles for mid frequencies
-        val triangleCount = 8
-        for (i in 0 until triangleCount) {
-            val angle = (i.toFloat() / triangleCount * 360f) * PI.toFloat() / 180f
-            val distance = 200f + mid * 100f
-            val size = 40f + mid * 60f
-            
-            val x = centerX + cos(angle) * distance
-            val y = centerY + sin(angle) * distance
-            
-            val path = Path().apply {
-                moveTo(x, y - size / 2)
-                lineTo(x + size / 2, y + size / 2)
-                lineTo(x - size / 2, y + size / 2)
-                close()
-            }
-            
-            drawPath(
-                path = path,
-                color = secondaryColor.copy(alpha = 0.6f + mid * 0.4f)
-            )
-        }
-        
-        // Small circles for treble
-        val circleCount = 16
-        for (i in 0 until circleCount) {
-            val angle = (i.toFloat() / circleCount * 360f) * PI.toFloat() / 180f
-            val distance = 100f
-            val x = centerX + cos(angle) * distance
-            val y = centerY + sin(angle) * distance
-            
-            drawCircle(
-                color = tertiaryColor.copy(alpha = treble),
-                radius = treble * 20f,
-                center = Offset(x, y)
-            )
-        }
-    }
-}
 
 /**
  * Kaleidoscope visualization
@@ -907,56 +400,6 @@ private fun KaleidoscopeVisualizer(
     }
 }
 
-/**
- * Fractal-inspired visualization
- */
-@Composable
-private fun FractalVisualizer(
-    state: VisualizerState,
-    rotation: Float,
-    primaryColor: Color,
-    tertiaryColor: Color
-) {
-    val bass = state.frequencyBands.bass
-    val mid = state.frequencyBands.mid
-    val treble = state.frequencyBands.treble
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        
-        // Recursive circle pattern
-        fun drawFractalCircle(x: Float, y: Float, radius: Float, depth: Int, angle: Float) {
-            if (depth <= 0 || radius < 5f) return
-            
-            val intensity = when (depth) {
-                3 -> bass
-                2 -> mid
-                else -> treble
-            }
-            
-            drawCircle(
-                color = androidx.compose.ui.graphics.lerp(primaryColor, tertiaryColor, depth / 3f)
-                    .copy(alpha = 0.3f + intensity * 0.7f),
-                radius = radius,
-                center = Offset(x, y),
-                style = Stroke(width = 2f)
-            )
-            
-            // Draw child circles
-            val childCount = 4
-            val childRadius = radius * 0.4f
-            for (i in 0 until childCount) {
-                val childAngle = angle + (i * 360f / childCount) * PI.toFloat() / 180f
-                val childX = x + cos(childAngle) * radius * 0.6f
-                val childY = y + sin(childAngle) * radius * 0.6f
-                drawFractalCircle(childX, childY, childRadius, depth - 1, angle + rotation)
-            }
-        }
-        
-        drawFractalCircle(centerX, centerY, min(size.width, size.height) * 0.25f, 3, rotation)
-    }
-}
 
 /**
  * Matrix rain style visualization
@@ -1001,154 +444,7 @@ private fun MatrixRainVisualizer(
     }
 }
 
-/**
- * Dual-channel stereo visualization
- */
-@Composable
-private fun DualChannelVisualizer(
-    state: VisualizerState,
-    primaryColor: Color,
-    secondaryColor: Color
-) {
-    val waveform = state.waveform.ifEmpty { List(256) { 0f } }
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerY = size.height / 2
-        val stepX = size.width / waveform.size
-        val amplitude = size.height * 0.2f
-        
-        // Left channel (top half)
-        val leftPath = Path().apply {
-            moveTo(0f, centerY / 2)
-            waveform.forEachIndexed { index, value ->
-                val x = index * stepX
-                val y = centerY / 2 + (value * amplitude * 0.8f)
-                lineTo(x, y)
-            }
-        }
-        
-        // Right channel (bottom half) - slightly offset
-        val rightPath = Path().apply {
-            moveTo(0f, centerY + centerY / 2)
-            waveform.forEachIndexed { index, value ->
-                val x = index * stepX
-                // Phase shift for stereo effect
-                val shiftedIdx = (index + waveform.size / 16) % waveform.size
-                val y = centerY + centerY / 2 + (waveform[shiftedIdx] * amplitude * 0.8f)
-                lineTo(x, y)
-            }
-        }
-        
-        // Draw center divider
-        drawLine(
-            color = primaryColor.copy(alpha = 0.2f),
-            start = Offset(0f, centerY),
-            end = Offset(size.width, centerY),
-            strokeWidth = 1f
-        )
-        
-        drawPath(
-            path = leftPath,
-            color = primaryColor,
-            style = Stroke(width = 2f, cap = StrokeCap.Round)
-        )
-        
-        drawPath(
-            path = rightPath,
-            color = secondaryColor,
-            style = Stroke(width = 2f, cap = StrokeCap.Round)
-        )
-    }
-}
 
-/**
- * 3D Cube spectrum visualization
- */
-@Composable
-private fun Cube3DVisualizer(
-    state: VisualizerState,
-    rotation: Float,
-    primaryColor: Color,
-    secondaryColor: Color,
-    tertiaryColor: Color
-) {
-    val spectrum = state.frequencyBands.spectrum.ifEmpty { List(64) { 0f } }
-    
-    Canvas(modifier = Modifier.fillMaxSize()) {
-        val centerX = size.width / 2
-        val centerY = size.height / 2
-        val cubeSize = min(size.width, size.height) * 0.3f
-        
-        // Pseudo-3D rotation
-        val rotX = rotation * PI.toFloat() / 180f
-        val rotY = (rotation * 1.3f) * PI.toFloat() / 180f
-        
-        // Draw cube edges with spectrum heights
-        val edges = listOf(
-            // Front face
-            listOf(-1f to -1f, 1f to -1f),
-            listOf(1f to -1f, 1f to 1f),
-            listOf(1f to 1f, -1f to 1f),
-            listOf(-1f to 1f, -1f to -1f),
-            // Back face
-            listOf(-0.7f to -0.7f, 0.7f to -0.7f),
-            listOf(0.7f to -0.7f, 0.7f to 0.7f),
-            listOf(0.7f to 0.7f, -0.7f to 0.7f),
-            listOf(-0.7f to 0.7f, -0.7f to -0.7f)
-        )
-        
-        edges.forEachIndexed { index, (start, end) ->
-            val magnitude = spectrum.getOrElse(index % spectrum.size) { 0f }
-            val lineColor = when (index % 3) {
-                0 -> primaryColor
-                1 -> secondaryColor
-                else -> tertiaryColor
-            }.copy(alpha = 0.5f + magnitude * 0.5f)
-            
-            // Simple 2D projection
-            val (x1, y1) = start
-            val (x2, y2) = end
-            val scale = cubeSize * (1f + magnitude * 0.5f)
-            
-            drawLine(
-                color = lineColor,
-                start = Offset(
-                    centerX + x1 * scale * cos(rotY),
-                    centerY + y1 * scale
-                ),
-                end = Offset(
-                    centerX + x2 * scale * cos(rotY),
-                    centerY + y2 * scale
-                ),
-                strokeWidth = 3f + magnitude * 3f,
-                cap = StrokeCap.Round
-            )
-        }
-        
-        // Draw spectrum bars on cube faces
-        spectrum.take(12).forEachIndexed { index, magnitude ->
-            val angle = (index * 30f + rotation) * PI.toFloat() / 180f
-            val radius = cubeSize
-            val height = magnitude * 100f
-            
-            val x = centerX + cos(angle) * radius
-            val y = centerY + sin(angle) * radius
-            
-            drawLine(
-                brush = Brush.verticalGradient(
-                    colors = listOf(
-                        primaryColor.copy(alpha = magnitude),
-                        tertiaryColor.copy(alpha = magnitude * 0.3f)
-                    )
-                ),
-                start = Offset(x, y),
-                end = Offset(x, y - height),
-                strokeWidth = 6f,
-                cap = StrokeCap.Round
-            )
-        }
-    }
-}
 
 /**
  * Available visualizer styles
